refactor(i18n): apply PR #227 review fixes
- Project `displayName` via a new `localizedStrict()` GROQ helper that
  does NOT coalesce missing slots. The default `localized()` would
  cross-fill an editor's en-only entry into ko/zh, defeating
  `pickLocalized`'s fallback to `name`/`title`.
- Standardise `Locale` imports on `@/i18n/routing` (the truth source);
  the three resource pages that still imported via `@/lib/content/home`
  now match the helper and the datasheets page.
- Reuse the exported `LocalizedField` type in page-local item shapes
  (datasheet/manual/drawing/cert) instead of inlining the
  `{ko?,en?,zh?}` triplet four times; cert also picks up the dedup for
  `issuer`/`scope`.
